% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_ckd.R
\name{curves_ckd}
\alias{curves_ckd}
\title{Build ROC/PR Curves from Probabilities}
\usage{
curves_ckd(df, prob_test)
}
\arguments{
\item{df}{A data.frame. Provide cleaned testing data with column
\code{classification}.}

\item{prob_test}{A data.frame. Provide two probability columns
\code{notckd} and \code{ckd}.}
}
\value{
A \code{precrec} object. Provide curve data usable by \code{ggplot2::autoplot}
and metrics obtainable via \code{precrec::auc}.
}
\description{
This function constructs ROC and PRC data for binary classification
using \code{precrec}, given true labels and predicted probabilities.
}
\examples{
\dontrun{
# just an example, this function is not exported by design.
pr  <- predict_ckd(test_df, glm_fit)
evc <- curves_ckd(test_df, pr)
ggplot2::autoplot(evc)
precrec::auc(evc)
}

}
\seealso{
\code{model_evaluation_ckd()}

Other ckd-evaluation-helpers-internal: 
\code{\link{metrics_ckd}()},
\code{\link{threshold_optimization_ckd}()}
}
\concept{ckd-evaluation-helpers-internal}
\keyword{internal}
