% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_ckd.R
\name{data_process_ckd}
\alias{data_process_ckd}
\title{Data Clean and Process for Chronic Kidney Disease (ckd) Dataset}
\usage{
data_process_ckd(file_name, row_threshold = 0.3, train_percentage = 0.7)
}
\arguments{
\item{file_name}{A character string. Path to raw CSV file.}

\item{row_threshold}{A numeric number. Row drop threshold for missing rate
(0-1), rows with missing rate over threshold are dropped.}

\item{train_percentage}{A numeric number. Train fraction in (0,1),
this percentage of random rows are extracted as train-dataset.}
}
\value{
A list. With following elements:
\itemize{
\item train_df            A data.frame. Cleaned training data.
\item test_df             A data.frame. Cleaned test data.
\item row_missing_table   A table. Per-row missing level table.
\item col_missing_table   A table. Per-col missing level table.
}
}
\description{
This function conducts manipulations to a raw dataset, including:
\enumerate{
\item Filling in empty and mess cells with NA;
\item Dropping rows with too much missing values;
\item Splitting dataset into train-dataset and test-dataset;
\item Handling with NA with different strategies;
Return tidy objects for downstream modeling.
}
}
\examples{
set.seed(123)
res <- data_process_ckd(
  system.file("extdata", "kidney_disease.csv", package="ckdPrediction"),
  row_threshold = 0.3,
  train_percentage = 0.7
  )
head(res$train_df, 3)
head(res$test_df, 3)
head(res$row_missing_table, 3)
head(res$col_missing_table, 3)

}
\seealso{
data_split_ckd(), row_process_ckd(), col_process_ckd(),
common_process_ckd()

Other ckd-prediction-helpers: 
\code{\link{model_comparison_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
