% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_ckd.R
\name{data_split_ckd}
\alias{data_split_ckd}
\title{Split Dataset into Training and Test Partitions}
\usage{
data_split_ckd(df, train_percentage = 0.7)
}
\arguments{
\item{df}{A data.frame. Provide the row-filtered dataset to be split.}

\item{train_percentage}{A numeric number. Specify the fraction in (0, 1) for
the training set; the remainder is assigned to the test set.}
}
\value{
A list. Return two data partitions for model development.
\itemize{
\item train_df          A data.frame. Training split of the input data.
\item test_df           A data.frame. Test split of the input data.
}
}
\description{
This function samples row indices to create training and test splits with a
user-defined training fraction.
}
\examples{
\dontrun{
# just an example, this function is not exported by design.
sp <- data_split_ckd(df = filtered, train_percentage = 0.7)
str(sp$train_df)
str(sp$test_df)
}

}
\seealso{
\code{data_process_ckd()}

Other ckd-prediction-helpers-internal: 
\code{\link{calculate_shap_ckd}()},
\code{\link{col_process_ckd}()},
\code{\link{common_process_ckd}()},
\code{\link{row_process_ckd}()}
}
\concept{ckd-prediction-helpers-internal}
\keyword{internal}
