% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_ckd.R
\name{metrics_ckd}
\alias{metrics_ckd}
\title{Compute Confusion-Matrix Metrics at a Given Threshold}
\usage{
metrics_ckd(df, prob_test, threshold = 0.5)
}
\arguments{
\item{df}{A data.frame. Provide cleaned testing data with column
\code{classification}.}

\item{prob_test}{A data.frame. Provide two probability columns
\code{notckd} and \code{ckd}.}

\item{threshold}{A numeric number. Specify decision threshold for model
in (0, 1).}
}
\value{
A list. With following elements:
\itemize{
\item accuracy   A numeric number. Compute overall accuracy.
\item precision  A numeric number. Compute positive predictive value.
\item recall     A numeric number. Compute sensitivity (TPR).
\item fpr        A numeric number. Compute false positive rate.
\item cm         A table. Provide confusion matrix with
levels \code{notckd}/\code{ckd}.
}
}
\description{
This function computes accuracy, precision, recall,
false positive rate (FPR), and return a confusion matrix given
class probabilities and a decision threshold.
}
\examples{
\dontrun{
# just an example, this function is not exported by design.
pr <- predict_ckd(test_df, rf_fit)
out <- metrics_ckd(test_df, pr, threshold = 0.5)
out$cm; out$accuracy
}

}
\seealso{
\code{model_evaluation_ckd()}

Other ckd-evaluation-helpers-internal: 
\code{\link{curves_ckd}()},
\code{\link{threshold_optimization_ckd}()}
}
\concept{ckd-evaluation-helpers-internal}
\keyword{internal}
