% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_ckd.R
\name{model_comparison_ckd}
\alias{model_comparison_ckd}
\title{Compare Performance Metrics of Multiple Models on Same Testing Set}
\usage{
model_comparison_ckd(model_list, df, threshold = 0.5)
}
\arguments{
\item{model_list}{A list. Contain trained models from \code{model_train_ckd()}.}

\item{df}{A data.frame. Cleaned testing data returned by
\code{data_process_ckd()}.}

\item{threshold}{A numeric number. Specify decision threshold for model
in (0, 1).}
}
\value{
A ggplot object. A bar chart to comparison performances among models.
}
\description{
This function generates a bar plot comparing metrics among models including:
\enumerate{
\item Accuracy;
\item Precision;
\item Recall;
\item Specificity;
\item F1-score;
\item AUC ROC;
\item AUC PRC;
under a same threshold for prediction classification.
In order to compare model performance under a desired threshold.
}
}
\examples{
set.seed(123)
test_df <- readRDS(
  system.file("extdata", "test_df.rds", package="ckdPrediction")
  )
rf_fit <- readRDS(
  system.file("extdata", "rf_fit.rds", package="ckdPrediction")
  )
glm_fit <- readRDS(
  system.file("extdata", "glm_fit.rds", package="ckdPrediction")
)
svm_fit <- readRDS(
  system.file("extdata", "svm_fit.rds", package="ckdPrediction")
)
model_list <- list(rf_fit, glm_fit, svm_fit)
plot <- model_comparison_ckd(model_list, test_df)
plot

}
\seealso{
model_evaluation_ckd(), predict_ckd()

Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
