% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_ckd.R
\name{model_evaluation_ckd}
\alias{model_evaluation_ckd}
\title{Full Evaluation for Single Model on Given Dataset}
\usage{
model_evaluation_ckd(df, model, threshold = 0.5)
}
\arguments{
\item{df}{A data.frame. Cleaned testing data returned by
\code{data_process_ckd()}.}

\item{model}{A list containing model returned by \code{model_train_ckd()}.}

\item{threshold}{Numeric threshold used for model evaluation, (0,1).}
}
\value{
A list. With following elements:
\itemize{
\item metrics         A list. Contain accuracy; precision; recall; fpr;
and a confusion-matrix
\item curves          A list. Contain a curve's plot and metrics
\item best_threshold  A list. Contain info and value of
a recommended threshold
}
}
\description{
This function predicts the data using a given model,
and evaluating the model at a certain threshold for classification.
Metrics including:
\enumerate{
\item accuracy;
\item recall;
\item false positive rate (FPR);
\item confusion matrix;
\item PRC curve with its metrics, such as area under curve (AUC);
\item ROC curve with its metrics, such as area under curve (AUC);
\item a recommended threshold, decided by highest F1-Score.
}
}
\examples{
set.seed(123)
test_df <- readRDS(
  system.file("extdata", "test_df.rds", package="ckdPrediction")
  )
svm_fit <- readRDS(
  system.file("extdata", "svm_fit.rds", package="ckdPrediction")
  )
ev <- model_evaluation_ckd(test_df, svm_fit, threshold = 0.5)
ev$metrics$accuracy
ev$metrics$cm
ev$curves$plot
ev$curves$auc
ev$best_threshold$info
ev$best_threshold$value

}
\seealso{
predict_ckd(), model_train_ckd(), metrics_ckd(), curves_ckd(),
threshold_optimization_ckd()

Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_comparison_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
