% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_ckd.R
\name{model_interpretation_importance_ckd}
\alias{model_interpretation_importance_ckd}
\title{Plot Feature Importance for rf / glm / svm Models}
\usage{
model_interpretation_importance_ckd(df, model)
}
\arguments{
\item{df}{A data.frame. A dataset returned by \code{data_process_ckd()},
(training/testing/full data are all available).}

\item{model}{A list. Contain model returned by \code{model_train_ckd()}.}
}
\value{
A ggplot2 object. A feature importance plot.
}
\description{
This function computes and visualizes feature importance for models.
Each model uses specific interpretation method for feature importance:
\enumerate{
\item Random Forest: MeanDecreaseGini;
\item (Ridge) Logistic Regression: |beta_j| * sd(x_j)$
\item Support Vector Machine: permutation-based delta-Accuracy.
}
}
\examples{
set.seed(123)
train_df <- readRDS(
  system.file("extdata", "train_df.rds", package="ckdPrediction")
  )
svm_fit <- readRDS(
  system.file("extdata", "svm_fit.rds", package="ckdPrediction")
  )
rf_importance_plot <- model_interpretation_importance_ckd(train_df, svm_fit)
rf_importance_plot

}
\seealso{
model_train_ckd()

Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_comparison_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
