% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_ckd.R
\name{model_interpretation_shap_ckd}
\alias{model_interpretation_shap_ckd}
\title{Plot SHAP Explanations for Models}
\usage{
model_interpretation_shap_ckd(
  train_df,
  test_df,
  model,
  visualization_type = 1,
  visualization_feature = "sc",
  visualization_row = 1
)
}
\arguments{
\item{train_df}{A data frame. Cleaned training data returned by
\code{data_process_ckd()}.}

\item{test_df}{A data frame. Cleaned testing data returned by
\code{data_process_ckd()}.}

\item{model}{A list. Contain model returned by \code{model_train_ckd()}.}

\item{visualization_type}{A numeric number. Represent the plot type:
type = 1: SHAP plot (beeswarm) for all features and all samples;
type = 2: SHAP plot (dependence) for one specific features and all samples
(need input of \code{visualization_feature});
type = 3: SHAP plot (waterfall) for all features and one specific sample
(need input of \code{visualization_row}).}

\item{visualization_feature}{A numeric number. Represent a single column
drawn in the plot.}

\item{visualization_row}{A numeric number. Represent a single row id drawn
in the plot.}
}
\value{
A ggplot2 object. A SHAP plot.
}
\description{
This function calculates SHAP values, and return a plot
in one of the three types:
type = 1: SHAP plot for all features and all samples;
type = 2: SHAP plot for one specific features and all samples
(need input of \code{visualization_feature});
type = 3: SHAP plot for all features and one specific sample
(need input of \code{visualization_row}).
}
\examples{
# This example is time & resource consuming, please wait for a short while.
set.seed(123)
train_df <- readRDS(
  system.file("extdata", "train_df.rds", package="ckdPrediction")
  )
test_df <- readRDS(
  system.file("extdata", "test_df.rds", package="ckdPrediction")
  )
svm_fit <- readRDS(
  system.file("extdata", "svm_fit.rds", package="ckdPrediction")
  )
plot_beeswarm <- model_interpretation_shap_ckd(
  train_df, test_df, svm_fit, visualization_type = 1
)
plot_beeswarm

plot_dependence <- model_interpretation_shap_ckd(
  train_df, test_df, svm_fit, visualization_type = 2,
  visualization_feature = "sc"
)
plot_dependence

plot_waterfall <- model_interpretation_shap_ckd(
  train_df, test_df, svm_fit, visualization_type = 3,
  visualization_row = 1
)
plot_waterfall

}
\seealso{
calculate_shap_ckd()

Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_comparison_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
