% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train_ckd.R
\name{model_train_ckd}
\alias{model_train_ckd}
\title{Train Machine Learning Models (rf/glm/svm)}
\usage{
model_train_ckd(df, type = 1, ...)
}
\arguments{
\item{df}{A data.frame. Cleaned training data returned by
\code{data_process_ckd()}.}

\item{type}{A numeric number. Representing the model type:
1 = random forest (rf);
2 = (ridge) logistic regression (glm);
3 = support vector machine (svm).}

\item{...}{Extra arguments passed to the underlying learner
(e.g., super-parameters for certain model supported by their packages).}
}
\value{
A list. With following elements:
\itemize{
\item model       A model object. Fitted model.
\item model_type  A character string. Name labels for the models,
one of "rf_fit","glm_fit","svm_fit"
}
}
\description{
This function uses cleaned ckd data to train one of the three ML-models:
\enumerate{
\item random forest (rf);
\item (ridge) logistic regression (glm);
\item support vector machine (svm).
}
}
\examples{
set.seed(123)
train_df <- readRDS(
  system.file("extdata", "train_df.rds", package="ckdPrediction")
  )
rf_fit <- model_train_ckd(train_df, type = 1)
rf_fit
glm_fit <- model_train_ckd(train_df, type = 2)
glm_fit
svm_fit <- model_train_ckd(train_df, type = 3)
svm_fit

}
\seealso{
Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_comparison_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{predict_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
