% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ckd.R
\name{predict_ckd}
\alias{predict_ckd}
\title{Predict Probabilities for Models}
\usage{
predict_ckd(df, model)
}
\arguments{
\item{df}{A data frame. Cleaned testing data returned by
\code{data_process_ckd()}.}

\item{model}{A list. Contain model returned by \code{model_train_ckd()}.}
}
\value{
A data.frame. With columns \code{notckd}, \code{ckd}.
}
\description{
The function eturns a two-column matrix/data.frame of class probabilities.
Used for evaluation and curves.
}
\examples{
set.seed(123)
test_df <- readRDS(
  system.file("extdata", "test_df.rds", package="ckdPrediction")
  )
svm_fit <- readRDS(
  system.file("extdata", "svm_fit.rds", package="ckdPrediction")
  )
predict_prob <- predict_ckd(test_df, svm_fit)
head(predict_prob)

}
\seealso{
predict_class_ckd(), model_evaluation_ckd(), model_train_ckd()

model_train_ckd()

Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_comparison_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_class_ckd}()}
}
\concept{ckd-prediction-helpers}
