% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ckd.R
\name{predict_class_ckd}
\alias{predict_class_ckd}
\title{Predict Class Labels for Models}
\usage{
predict_class_ckd(df, model, threshold = 0.5)
}
\arguments{
\item{df}{A data frame. Cleaned testing data returned by
\code{data_process_ckd()}.}

\item{model}{A list. Contain model returned by \code{model_train_ckd()}.}

\item{threshold}{A nemeric number in (0,1), default is 0.5.
The "ckd" probability over this threshold lead to a "ckd" prediction.}
}
\value{
A factor. Contain all predicted labels (\code{notckd} / \code{ckd}).
}
\description{
The function returns a two-column matrix/data.frame of class labels.
Used for evaluation and curves.
}
\examples{
set.seed(123)
test_df <- readRDS(
  system.file("extdata", "test_df.rds", package="ckdPrediction")
  )
svm_fit <- readRDS(
  system.file("extdata", "svm_fit.rds", package="ckdPrediction")
  )
predict_class <- predict_class_ckd(test_df, svm_fit, threshold = 0.5)
head(predict_class)

}
\seealso{
predict_ckd(), model_evaluation_ckd()

Other ckd-prediction-helpers: 
\code{\link{data_process_ckd}()},
\code{\link{model_comparison_ckd}()},
\code{\link{model_evaluation_ckd}()},
\code{\link{model_interpretation_importance_ckd}()},
\code{\link{model_interpretation_shap_ckd}()},
\code{\link{model_train_ckd}()},
\code{\link{predict_ckd}()}
}
\concept{ckd-prediction-helpers}
